requirements.txt
acres==0.3.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.1
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
configobj==5.0.9
configparser==7.2.0
cryptography==45.0.3
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
emoji==2.14.1
etelemetry==0.3.1
fastapi==0.115.12
ffmpy==0.6.0
filelock==3.18.0
filetype==1.2.0
frozenlist==1.6.0
fsspec==2025.5.1
gradio==5.32.1
gradio_client==1.10.2
grandalf==0.8
groovy==0.1.2
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.2
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.3
idna==3.10
iniconfig==2.1.0
isodate==0.6.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.61
langchain-mcp-adapters==0.1.1
langchain-ollama==0.3.3
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
langdetect==1.0.9
langgraph==0.4.7
langgraph-checkpoint==2.0.26
langgraph-checkpoint-sqlite==2.0.10
langgraph-prebuilt==0.2.1
langgraph-sdk==0.1.70
langsmith==0.3.42
looseversion==1.3.0
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.9.1
mdurl==0.1.2
motor==3.7.1
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numpy==2.2.6
olefile==0.47
ollama==0.4.8
openai==1.82.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pdf2image==1.17.0
pillow==11.2.1
pip==25.1.1
pluggy==1.6.0
propcache==0.3.1
prov==2.0.1
psutil==7.0.0
puremagic==1.29
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydantic-settings==2.9.1
pydot==4.0.0
pydub==0.25.1
Pygments==2.19.1
pymongo==4.13.0
PyMuPDF==1.26.0
pyparsing==3.2.3
pypdf==5.5.0
pytesseract==0.3.13
pytest==8.3.5
pytest-asyncio==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
RapidFuzz==3.13.0
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
ruff==0.11.12
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
semantic-version==2.10.0
sentence-transformers==5.0.0
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sqlite-vec==0.1.6
sse-starlette==2.3.5
starlette==0.46.2
sympy==1.14.0
tavily-python==0.7.3
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
tomlkit==0.13.2
torch==2.7.1
tqdm==4.67.1
traits==7.0.2
transformers==4.53.1
typer==0.16.0
typing_extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
unstructured==0.17.2
unstructured-client==0.35.0
urllib3==2.4.0
uvicorn==0.34.2
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.2
XlsxWriter==3.2.3
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0